Refactor trainer form components and logic
- Introduced a custom hook `useTrainerForm` to encapsulate form state and logic.
- Created separate components for each step of the trainer form: `UserDataStep`, `PersonalDataStep`, and `WorkDataStep`.
- Added `FormHeader` and `FormNavigation` components for better structure and navigation.
- Moved validation schemas and constants to a dedicated file for better organization.
- Updated the main `NewTrainerForm` component to utilize the new structure and components.
- Improved error handling and validation feedback throughout the form.
