+++ apps/web/src/lib/trainer-types.ts
@@ -1,0 +1,5 @@
+// ===== TIPOS PARA ENTRENADORES =====
+
+export interface FormErrors {
+    [key: string]: string;
+}
